Modify an instruction associated with how to activate LPK option
Even though you build an openssh package with this lpk patch according
to the README file in the first release, you will get the following
errors,
/etc/ssh/sshd_config: line 99: Bad configuration option: UseLPK
/etc/ssh/sshd_config: line 101: Bad configuration option: LpkServers
/etc/ssh/sshd_config: line 103: Bad configuration option: LpkUserDN
/etc/ssh/sshd_config: line 104: Bad configuration option: LpkGroupDN
/etc/ssh/sshd_config: line 107: Bad configuration option: LpkServerGroup
/etc/ssh/sshd_config: line 108: Bad configuration option: LpkForceTLS
/etc/ssh/sshd_config: line 110: Bad configuration option: LpkBindTimelimit
/etc/ssh/sshd_config: terminating, 7 bad configuration options
This is because LPK option is not activated in building the package.
You have to add a cpp flag, -DWITH_LDAP_PUBKEY and also a flag for
linker, -lldap.